Add Column ancestry tracking, Query wildcard tests
Column 'ancestors' or 'parents' are now tracked. When two tables are
joined, the resulting table contains a column that is the 'child' of the
join columns on the original two tables. Now that we track this
relationship, we can delay application of where conditions until after
the join even if the where condition explicitly specifies a parent
column instead of the child column.

Also, tests were added to QueryTest to verify that asterisk wildcards
are expanded to include all table columns.